# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Smart Attendance Management System API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type LoginInputRole = typeof LoginInputRole[keyof typeof LoginInputRole];


export const LoginInputRole = {
  admin: 'admin',
  faculty: 'faculty',
  student: 'student',
} as const;

export interface LoginInput {
  email: string;
  password: string;
  role: LoginInputRole;
}

export type AuthUserRole = typeof AuthUserRole[keyof typeof AuthUserRole];


export const AuthUserRole = {
  admin: 'admin',
  faculty: 'faculty',
  student: 'student',
} as const;

export interface AuthUser {
  id: number;
  name: string;
  email: string;
  role: AuthUserRole;
  /** @nullable */
  department?: string | null;
  /** @nullable */
  semester?: string | null;
}

export interface AuthResponse {
  token: string;
  user: AuthUser;
}

export interface Student {
  id: number;
  name: string;
  email: string;
  department: string;
  semester: string;
  createdAt: string;
}

export interface StudentInput {
  name: string;
  email: string;
  password: string;
  department: string;
  semester: string;
}

export interface StudentUpdate {
  name?: string;
  email?: string;
  department?: string;
  semester?: string;
  /** @nullable */
  password?: string | null;
}

export interface SubjectAttendanceSummary {
  subjectId: number;
  subjectName: string;
  subjectCode: string;
  totalClasses: number;
  present: number;
  absent: number;
  late: number;
  percentage: number;
}

export interface Faculty {
  id: number;
  name: string;
  email: string;
  createdAt: string;
}

export interface FacultyInput {
  name: string;
  email: string;
  password: string;
}

export interface FacultyUpdate {
  name?: string;
  email?: string;
  /** @nullable */
  password?: string | null;
}

export interface Subject {
  id: number;
  subjectName: string;
  subjectCode: string;
  department: string;
  semester: string;
  /** @nullable */
  facultyId?: number | null;
  /** @nullable */
  facultyName?: string | null;
  createdAt: string;
}

export interface SubjectInput {
  subjectName: string;
  subjectCode: string;
  department: string;
  semester: string;
  /** @nullable */
  facultyId?: number | null;
}

export interface SubjectUpdate {
  subjectName?: string;
  subjectCode?: string;
  department?: string;
  semester?: string;
  /** @nullable */
  facultyId?: number | null;
}

export type AttendanceRecordStatus = typeof AttendanceRecordStatus[keyof typeof AttendanceRecordStatus];


export const AttendanceRecordStatus = {
  Present: 'Present',
  Absent: 'Absent',
  Late: 'Late',
} as const;

export interface AttendanceRecord {
  id: number;
  studentId: number;
  studentName: string;
  subjectId: number;
  subjectName: string;
  /** @nullable */
  facultyId?: number | null;
  date: string;
  status: AttendanceRecordStatus;
  createdAt: string;
}

export type AttendanceEntryStatus = typeof AttendanceEntryStatus[keyof typeof AttendanceEntryStatus];


export const AttendanceEntryStatus = {
  Present: 'Present',
  Absent: 'Absent',
  Late: 'Late',
} as const;

export interface AttendanceEntry {
  studentId: number;
  status: AttendanceEntryStatus;
}

export interface AttendanceInput {
  subjectId: number;
  date: string;
  /** @nullable */
  facultyId?: number | null;
  entries: AttendanceEntry[];
}

export type AttendanceUpdateStatus = typeof AttendanceUpdateStatus[keyof typeof AttendanceUpdateStatus];


export const AttendanceUpdateStatus = {
  Present: 'Present',
  Absent: 'Absent',
  Late: 'Late',
} as const;

export interface AttendanceUpdate {
  status: AttendanceUpdateStatus;
}

export interface DashboardStats {
  totalStudents: number;
  totalFaculty: number;
  totalSubjects: number;
  overallAttendancePercentage: number;
  todayPresent: number;
  todayAbsent: number;
  totalAttendanceRecords: number;
}

export interface AttendanceTrendPoint {
  date: string;
  present: number;
  absent: number;
  late: number;
  total: number;
}

export interface LowAttendanceAlert {
  studentId: number;
  studentName: string;
  department: string;
  semester: string;
  subjectId: number;
  subjectName: string;
  percentage: number;
  totalClasses: number;
  present: number;
}

export interface SubjectStat {
  subjectId: number;
  subjectName: string;
  subjectCode: string;
  department: string;
  semester: string;
  totalClasses: number;
  averageAttendance: number;
}

export type ListStudentsParams = {
department?: string;
semester?: string;
search?: string;
};

export type ListFacultyParams = {
search?: string;
};

export type ListSubjectsParams = {
department?: string;
semester?: string;
facultyId?: number;
};

export type ListAttendanceParams = {
subjectId?: number;
studentId?: number;
date?: string;
dateFrom?: string;
dateTo?: string;
};

export type GetAttendanceTrendsParams = {
subjectId?: number;
period?: GetAttendanceTrendsPeriod;
dateFrom?: string;
dateTo?: string;
};

export type GetAttendanceTrendsPeriod = typeof GetAttendanceTrendsPeriod[keyof typeof GetAttendanceTrendsPeriod];


export const GetAttendanceTrendsPeriod = {
  weekly: 'weekly',
  monthly: 'monthly',
} as const;

export type GetLowAttendanceStudentsParams = {
threshold?: number;
subjectId?: number;
};