# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Smart Attendance Management System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login (student, faculty, or admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user profile
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUser"
        "401":
          description: Unauthorized

  # ─── STUDENTS ──────────────────────────────────────────────────────────────
...
  /attendance:
    get:
      operationId: listAttendance
      tags: [attendance]
      summary: List attendance records with filters
      parameters:
        - in: query
          name: subjectId
          schema:
            type: integer
        - in: query
          name: studentId
          schema:
            type: integer
        - in: query
          name: date
          schema:
            type: string
        - in: query
          name: dateFrom
          schema:
            type: string
        - in: query
          name: dateTo
          schema:
            type: string
      responses:
        "200":
          description: Attendance records
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AttendanceRecord"
    post:
      operationId: markAttendance
      tags: [attendance]
      summary: Mark attendance (bulk or single)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AttendanceInput"
      responses:
        "200":
          description: Attendance marked
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AttendanceRecord"
  /attendance/{id}:
    patch:
      operationId: updateAttendance
      tags: [attendance]
      summary: Update a single attendance record
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AttendanceUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AttendanceRecord"
        "404":
          description: Not found

  # ─── ANALYTICS ──────────────────────────────────────────────────────────────
...
  /analytics/attendance-trends:
    get:
      operationId: getAttendanceTrends
      tags: [analytics]
      summary: Get weekly/monthly attendance trends
      parameters:
        - in: query
          name: subjectId
          schema:
            type: integer
        - in: query
          name: period
          schema:
            type: string
            enum: [weekly, monthly]
        - in: query
          name: dateFrom
          schema:
            type: string
        - in: query
          name: dateTo
          schema:
            type: string
      responses:
        "200":
          description: Attendance trends
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AttendanceTrendPoint"
  /analytics/low-attendance:
    get:
      operationId: getLowAttendanceStudents
      tags: [analytics]
      summary: Get students with low attendance (below threshold)
      parameters:
        - in: query
          name: threshold
          schema:
            type: integer
        - in: query
          name: subjectId
          schema:
            type: integer
      responses:
        "200":
          description: Students with low attendance
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LowAttendanceAlert"
...
components:
  schemas:
...
    LoginInput:
      type: object
      required: [email, password, role]
      properties:
        email:
          type: string
        password:
          type: string
        role:
          type: string
          enum: [admin, faculty, student]
    AuthResponse:
      type: object
      required: [token, user]
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/AuthUser"
    AuthUser:
      type: object
      required: [id, name, email, role]
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [admin, faculty, student]
        department:
          type: ["string", "null"]
        semester:
          type: ["string", "null"]

    # ─── STUDENT ─────────────────────────────────────────────────────────────
...
    AttendanceRecord:
      type: object
      required: [id, studentId, studentName, subjectId, subjectName, date, status, createdAt]
      properties:
        id:
          type: integer
        studentId:
          type: integer
        studentName:
          type: string
        subjectId:
          type: integer
        subjectName:
          type: string
        facultyId:
          type: ["integer", "null"]
        date:
          type: string
        status:
          type: string
          enum: [Present, Absent, Late]
        createdAt:
          type: string
    AttendanceEntry:
      type: object
      required: [studentId, status]
      properties:
        studentId:
          type: integer
        status:
          type: string
          enum: [Present, Absent, Late]
    AttendanceInput:
      type: object
      required: [subjectId, date, entries]
      properties:
        subjectId:
          type: integer
        date:
          type: string
        facultyId:
          type: ["integer", "null"]
        entries:
          type: array
          items:
            $ref: "#/components/schemas/AttendanceEntry"
    AttendanceUpdate:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [Present, Absent, Late]

    # ─── ANALYTICS ───────────────────────────────────────────────────────────
...
    AttendanceTrendPoint:
      type: object
      required: [date, present, absent, late, total]
      properties:
        date:
          type: string
        present:
          type: integer
        absent:
          type: integer
        late:
          type: integer
        total:
          type: integer
    LowAttendanceAlert:
      type: object
      required: [studentId, studentName, department, semester, subjectId, subjectName, percentage, totalClasses, present]
      properties:
        studentId:
          type: integer
        studentName:
          type: string
        department:
          type: string
        semester:
          type: string
        subjectId:
          type: integer
        subjectName:
          type: string
        percentage:
          type: number
        totalClasses:
          type: integer
        present:
          type: integer
...