# lib/api-zod/src/generated/types/attendanceEntry.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Smart Attendance Management System API
 * OpenAPI spec version: 0.1.0
 */
import type { AttendanceEntryStatus } from './attendanceEntryStatus';

export interface AttendanceEntry {
  studentId: number;
  status: AttendanceEntryStatus;
}