# lib/api-zod/src/generated/types/attendanceTrendPoint.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Smart Attendance Management System API
 * OpenAPI spec version: 0.1.0
 */

export interface AttendanceTrendPoint {
  date: string;
  present: number;
  absent: number;
  late: number;
  total: number;
}